+++ wheel/python/chia_rs/chia_rs.pyi
@@ -1,18 +1,80 @@
-from typing import Optional, List, ClassVar, Any, Union, Tuple
+from typing import ClassVar, Any, Union, Tuple, Optional, List
 from .sized_bytes import bytes32
 
 
-class BLSCache:
-    def __init__(self, cache_size: Optional[int] = 50000) -> None: ...
-    def len(self) -> int: ...
-    def aggregate_verify(
-        self,
-        pks: List[G1Element],
-        msgs: List[bytes],
-        sig: G2Element,
-    ) -> bool: ...
-    def items(self) -> List[Tuple[bytes, bytes]]: ...
-    def update(self, other: List[Tuple[bytes, bytes]]) -> None: ...
+class AugSchemeMPL:
+    @staticmethod
+    def sign(sk: PrivateKey, msg: bytes, prepend_pk: Optional[G1Element] = None) -> G2Element: ...
+    @staticmethod
+    def aggregate(sigs: List[G2Element]) -> G2Element: ...
+    @staticmethod
+    def verify(pk: G1Element, msg: bytes, sig: G2Element) -> bool: ...
+    @staticmethod
+    def aggregate_verify(pks: List[G1Element], msgs: List[bytes], sig: G2Element) -> bool: ...
+    @staticmethod
+    def key_gen(seed: bytes) -> PrivateKey: ...
+    @staticmethod
+    def g2_from_message(msg: bytes) -> G2Element: ...
+    @staticmethod
+    def derive_child_sk(sk: PrivateKey, index: int) -> PrivateKey: ...
+    @staticmethod
+    def derive_child_sk_unhardened(sk: PrivateKey, index: int) -> PrivateKey: ...
+    @staticmethod
+    def derive_child_pk_unhardened(pk: G1Element, index: int) -> G1Element: ...
+
+class PrivateKey:
+    PRIVATE_KEY_SIZE: ClassVar[int] = ...
+    def sign_g2(self, msg: bytes, dst: bytes) -> G2Element: ...
+    def get_g1(self) -> G1Element: ...
+    def __str__(self) -> str: ...
+    def __init__(self) -> None: ...
+    def __hash__(self) -> int: ...
+    def __repr__(self) -> str: ...
+    def __richcmp__(self) -> Any: ...
+    def __deepcopy__(self) -> PrivateKey: ...
+    def __copy__(self) -> PrivateKey: ...
+    @staticmethod
+    def from_bytes(buffer: bytes) -> PrivateKey: ...
+    @staticmethod
+    def from_bytes_unchecked(buffer: bytes) -> PrivateKey: ...
+    @staticmethod
+    def parse_rust(buffer: ReadableBuffer, trusted: bool = False) -> Tuple[PrivateKey, int]: ...
+    def to_bytes(self) -> bytes: ...
+    def __bytes__(self) -> bytes: ...
+    def stream_to_bytes(self) -> bytes: ...
+    def get_hash(self) -> bytes32: ...
+    def to_json_dict(self) -> Any: ...
+    @staticmethod
+    def from_json_dict(json_dict: Any) -> PrivateKey: ...
+
+class G2Element:
+    SIZE: ClassVar[int] = ...
+    def __new__(cls) -> G2Element: ...
+    def pair(self, public_key: G1Element) -> GTElement: ...
+    @staticmethod
+    def generator() -> G2Element: ...
+    def __str__(self) -> str: ...
+    def __add__(self, rhs: G2Element) -> G2Element: ...
+    def __iadd__(self, rhs: G2Element) -> G2Element: ...
+    def __init__(self) -> None: ...
+    def __hash__(self) -> int: ...
+    def __repr__(self) -> str: ...
+    def __richcmp__(self) -> Any: ...
+    def __deepcopy__(self) -> G2Element: ...
+    def __copy__(self) -> G2Element: ...
+    @staticmethod
+    def from_bytes(buffer: bytes) -> G2Element: ...
+    @staticmethod
+    def from_bytes_unchecked(buffer: bytes) -> G2Element: ...
+    @staticmethod
+    def parse_rust(buffer: ReadableBuffer, trusted: bool = False) -> Tuple[G2Element, int]: ...
+    def to_bytes(self) -> bytes: ...
+    def __bytes__(self) -> bytes: ...
+    def stream_to_bytes(self) -> bytes: ...
+    def get_hash(self) -> bytes32: ...
+    def to_json_dict(self) -> Any: ...
+    @staticmethod
+    def from_json_dict(json_dict: Any) -> G2Element: ...
 
 class G1Element:
     SIZE: ClassVar[int] = ...
@@ -44,35 +106,6 @@ class G1Element:
     @staticmethod
     def from_json_dict(json_dict: Any) -> G1Element: ...
 
-class G2Element:
-    SIZE: ClassVar[int] = ...
-    def __new__(cls) -> G2Element: ...
-    def pair(self, public_key: G1Element) -> GTElement: ...
-    @staticmethod
-    def generator() -> G2Element: ...
-    def __str__(self) -> str: ...
-    def __add__(self, rhs: G2Element) -> G2Element: ...
-    def __iadd__(self, rhs: G2Element) -> G2Element: ...
-    def __init__(self) -> None: ...
-    def __hash__(self) -> int: ...
-    def __repr__(self) -> str: ...
-    def __richcmp__(self) -> Any: ...
-    def __deepcopy__(self) -> G2Element: ...
-    def __copy__(self) -> G2Element: ...
-    @staticmethod
-    def from_bytes(buffer: bytes) -> G2Element: ...
-    @staticmethod
-    def from_bytes_unchecked(buffer: bytes) -> G2Element: ...
-    @staticmethod
-    def parse_rust(buffer: ReadableBuffer, trusted: bool = False) -> Tuple[G2Element, int]: ...
-    def to_bytes(self) -> bytes: ...
-    def __bytes__(self) -> bytes: ...
-    def stream_to_bytes(self) -> bytes: ...
-    def get_hash(self) -> bytes32: ...
-    def to_json_dict(self) -> Any: ...
-    @staticmethod
-    def from_json_dict(json_dict: Any) -> G2Element: ...
-
 class GTElement:
     SIZE: ClassVar[int] = ...
     def __str__(self) -> str: ...
@@ -100,27 +133,14 @@ class GTElement:
 
 ReadableBuffer = Union[bytes, bytearray, memoryview]
 
-class PrivateKey:
-    PRIVATE_KEY_SIZE: ClassVar[int] = ...
-    def sign_g2(self, msg: bytes, dst: bytes) -> G2Element: ...
-    def get_g1(self) -> G1Element: ...
-    def __str__(self) -> str: ...
-    def __init__(self) -> None: ...
-    def __hash__(self) -> int: ...
-    def __repr__(self) -> str: ...
-    def __richcmp__(self) -> Any: ...
-    def __deepcopy__(self) -> PrivateKey: ...
-    def __copy__(self) -> PrivateKey: ...
-    @staticmethod
-    def from_bytes(buffer: bytes) -> PrivateKey: ...
-    @staticmethod
-    def from_bytes_unchecked(buffer: bytes) -> PrivateKey: ...
-    @staticmethod
-    def parse_rust(buffer: ReadableBuffer, trusted: bool = False) -> Tuple[PrivateKey, int]: ...
-    def to_bytes(self) -> bytes: ...
-    def __bytes__(self) -> bytes: ...
-    def stream_to_bytes(self) -> bytes: ...
-    def get_hash(self) -> bytes32: ...
-    def to_json_dict(self) -> Any: ...
-    @staticmethod
-    def from_json_dict(json_dict: Any) -> PrivateKey: ...
+class BLSCache:
+    def __init__(self, cache_size: Optional[int] = 50000) -> None: ...
+    def len(self) -> int: ...
+    def aggregate_verify(
+        self,
+        pks: List[G1Element],
+        msgs: List[bytes],
+        sig: G2Element,
+    ) -> bool: ...
+    def items(self) -> List[Tuple[bytes, bytes]]: ...
+    def update(self, other: List[Tuple[bytes, bytes]]) -> None: ...
